pyproject.toml
[project]
name = "bioinfo-sidecar"
version = "0.1.0"
description = "Local-first AI research assistant for bioinformatics — arXiv + Ollama/Qwen + pgvector + LangGraph + Gradio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.1.0",
    "docling>=2.0.0",
    "gradio>=4.0.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "marker-pdf>=0.3.0",
    "pgvector>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic-settings>=2.5.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "ipykernel>=6.29.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "ipykernel>=6.29.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
